requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2020.4.5.1
Click==7.0
colorama==0.4.3
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
